workflow-builder-be/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-genai==0.3.0
chromadb==0.4.22
sentence-transformers>=3.0.1
huggingface-hub>=0.23.0
pymupdf==1.23.15
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
numpy<2.0
httpx==0.27.2
torch --index-url https://download.pytorch.org/whl/cpu